+++ changelog.md
@@ -11,6 +11,10 @@ Older versions can be found in the documentation changelogs.
         * Added `pinned` and `unpinned` events
         * Added a `pinned` boolean property
         * Added an `unpin()` method that resets state
+        * Reworked the `elementTop` property to be the top offset value on page load and added `initialTop` to be the CSS defined top value
+        * Fixed a bug where `initialTop` would be set to `NaN`
+        * Fixed a bug where elements to be pinned were at the very top would pin immediately
+        * Fixed a bug regarding scroll top and `fixed` elements
 
 ## 2.1.5 ##
 * Added docs for integration with build tools and task runners