+++ changelog.md
@@ -2,6 +2,12 @@
 
 Older versions can be found in the documentation changelogs.
 
+## 2.1.7 ##
+* Updated the `.span-*` classes to no longer be coupled with the grid component
+* Components
+    * Stalker
+        * Fixed a bug where the bottom threshold was being calculated incorrectly
+
 ## 2.1.6 ##
 * Further improved AMD loading by splitting child components into their own files
 * Components